Error checking for +giveaway

+++ commands/giveaway.js
@@ -5,9 +5,16 @@ module.exports = {
     usage: "+giveaway <messageID> [winners]",
     roleWhitelist: ["620274909700161556"], //*
     async execute(message, strings, args) {
+            if (!args[0]) return message.channel.send("You forgot to specify a message to look for! Use the message ID")
+            const giveawayMsg = await message.channel.messages.fetch(args[0])
+            .catch(err => {
+                return message.channel.send("Couldn't find that message! Here's the error:\n" + err)
+            })
             message.delete()
-            const m = await message.channel.messages.fetch(args[0])
-            const users = await m.reactions.cache.find(r => r.emoji.name == "🎉").users.fetch()
+            const users = await giveawayMsg.reactions.cache.find(r => r.emoji.name == "🎉").users.fetch()
+            .catch(err => {
+                return message.channel.send("That message doesn't have any 🎉 reactions. Here's the error:\n" + err)
+            })
             const winner = users.random(args[1] || 1)
             let winners = []
             winner.forEach(user => winners.push(user.id))
